Crate jsonrpc_pubsub

Publish-Subscribe extension for JSON-RPC

Modules

manager

The SubscriptionManager used to manage subscription based RPCs.

oneshot

A futures oneshot channel that can be used for rendezvous.

typed

PUB-SUB auto-serializing structures.

Structs

IoDelegate

Wire up rpc subscriptions to delegate struct

PubSubHandler

Publish-Subscribe extension of IoHandler.

Session

RPC client session Keeps track of active subscriptions and unsubscribes from them upon dropping.

Sink

A handle to send notifications directly to subscribed client.

Subscriber

Represents a subscribing client. Subscription handlers can either reject this subscription request or assign an unique id.

Enums

SubscriptionId

Unique subscription id.

Traits

PubSubMetadata

Metadata extension for pub-sub method handling.

SubscribeRpcMethod

Subscribe handler

UnsubscribeRpcMethod

Unsubscribe handler

Functions

new_subscription

Creates new subscribe and unsubscribe RPC methods

Type Definitions

SinkResult

Subscription send result.

TransportError

Raw transport error.