Module jsonrpc_pubsub::manager

The SubscriptionManager used to manage subscription based RPCs.

The manager provides four main things in terms of functionality:

The ability to create unique subscription IDs through the use of the IdProvider trait. Two implementations are availble out of the box, a NumericIdProvider and a RandomStringIdProvider.

An executor with which to drive Futures to completion.

A way to add new subscriptions. Subscriptions should come in the form of a Stream. These subscriptions will be transformed into notifications by the manager, which can be consumed by the client.

A way to cancel any currently active subscription.

Structs

NumericIdProvider
Provides a thread-safe incrementing integer which can be used as a subscription ID.

RandomStringIdProvider
Used to generate random strings for use as subscription IDs.

SubscriptionManager
Subscriptions manager.

Traits

IdProvider
Trait used to provide unique subscription IDs.

Type Definitions

TaskExecutor
Alias for an implementation of futures::future::Executor.