Struct jsonrpc_pubsub::oneshot::Sender

pub struct Sender<T> { /* fields omitted */ }

A sender part of the channel.

Methods

impl<T> Sender<T>

pub fn send(self, t: T) -> Result<(), T>

Consume the sender and queue up an item to send.

This method returns right away and never blocks, there is no guarantee though that the message is received by the other end.

pub fn send_and_wait(self, t: T) -> impl Future<Item = (), Error = ()>

Consume the sender and send an item.

The returned future will resolve when the message is received on the other end. Note that polling the future is actually not required to send the message as that happens synchronously. The future resolves to error in case the receiving end was dropped before being able to process the message.