Trait jsonrpc_pubsub::PubSubMetadata

pub trait PubSubMetadata: Metadata {
    fn session(&self) -> Option<Arc<Session>>;
}

Metadata extension for pub-sub method handling.

NOTE storing PubSubMetadata (or rather storing Arc<Session>) in any other place outside of the handler will prevent unsubscribe methods to be called in case the Session is dropped (i.e. transport connection is closed).

Required methods

fn session(&self) -> Option<Arc<Session>>

Returns session object associated with given request/client. None indicates that sessions are not supported on the used transport.